small mgf parsing changes
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@897 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ branches/release_2015024/CometSearch/CometWritePepXML.cpp
@@ -435,7 +435,7 @@ void CometWritePepXML::PrintResults(int iWhichQuery,
             || strchr(pQuery->_spectrumInfoInternal.szNativeID, '>'))
       {
          fprintf(fpout, " spectrumNativeID=\"");
-         for (i=0; i<strlen(pQuery->_spectrumInfoInternal.szNativeID); i++)
+         for (i=0; i<(int)strlen(pQuery->_spectrumInfoInternal.szNativeID); i++)
          {
             switch(pQuery->_spectrumInfoInternal.szNativeID[i])
             {
